archive_deletewhenready/whispercpp_arm/requirements.txt
# Required deps for ARM whisper.cpp GUI with punctuation
python-docx>=1.1.0
numpy>=1.24
onnxruntime-directml>=1.18.0
